# Fix fill forward resolution adjustment on removal