--- a/Algorithm.CSharp/AllShortableSymbolsCoarseSelectionRegressionAlgorithm.cs
+++ b/Algorithm.CSharp/AllShortableSymbolsCoarseSelectionRegressionAlgorithm.cs
@@ -238,7 +238,7 @@ namespace QuantConnect.Algorithm.CSharp
         /// <summary>
         /// Data Points count of all timeslices of algorithm
         /// </summary>
-        public long DataPoints => 37754;
+        public long DataPoints => 37751;
 
         /// <summary>
         /// Data Points count of the algorithm history