--- a/Engine/DataFeeds/SubscriptionCollection.cs
+++ b/Engine/DataFeeds/SubscriptionCollection.cs
@@ -102,7 +102,12 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         {
             if (_subscriptions.TryRemove(configuration, out subscription))
             {
-                UpdateFillForwardResolution(FillForwardResolutionOperation.AfterRemove, configuration);
+                // for user friendlyness only look at removals triggerd by the user not those that are due to a data feed ending because of no more data,
+                // let's try to respect the users original FF enumerator request
+                if (!subscription.EndOfStream)
+                {
+                    UpdateFillForwardResolution(FillForwardResolutionOperation.AfterRemove, configuration);
+                }
                 _sortingSubscriptionRequired = true;
                 return true;
             }
@@ -179,7 +184,8 @@ namespace QuantConnect.Lean.Engine.DataFeeds
                 ||
                     (operation == FillForwardResolutionOperation.AfterRemove // We are removing
                     && configuration.Increment == _fillForwardResolution.Value // True: We are removing the resolution we were using
-                    && _subscriptions.All(x => x.Key.Resolution != configuration.Resolution))) // False: there is at least another one equal, no need to update
+                    // False: there is at least another one equal, no need to update, but we only look at those valid configuration which are the ones which set the FF resolution
+                    && _subscriptions.Keys.All(x => !ValidateFillForwardResolution(x) || x.Resolution != configuration.Resolution)))
                 )
             {
                 var configurations = (operation == FillForwardResolutionOperation.BeforeAdd)