--- a/Engine/DataFeeds/SubscriptionSynchronizer.cs
+++ b/Engine/DataFeeds/SubscriptionSynchronizer.cs
@@ -223,7 +223,9 @@ namespace QuantConnect.Lean.Engine.DataFeeds
                         // time pulse to align algorithm time with current frontier
                         yield return _timeSliceFactory.CreateTimePulse(frontierUtc);
 
-                        foreach (var kvp in universeData)
+                        // trigger the smalled resolution first, so that FF res get's set once from the start correctly
+                        // while at it, let's make it determininstic and sort by universe sid later
+                        foreach (var kvp in universeData.OrderBy(x => x.Key.Configuration.Resolution).ThenBy(x => x.Key.Symbol.ID))
                         {
                             var universe = kvp.Key;
                             var baseDataCollection = kvp.Value;