<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Fix fill forward resolution adjustment on subscription removal. Adding regression algorithm reproducing issue


`Algorithm.CSharp/RegressionAlgorithm.cs`: Stats changed because it would trigger a FF resolution change, which is no longer the case
> DataManager(): Fill forward resolution has changed from 00:01:00 to 1.00:00:00 at utc: 10/11/2013 7:59:59 PM. Restarting 1 subscriptions...

`Algorithm.CSharp/AllShortableSymbolsCoarseSelectionRegressionAlgorithm.cs`: Reproduced the issue where on FF resolution change the subscription wouldn't change as expected

**Performance impact running algorithm in issue**
```
MASTER
193.02 seconds at 535k data points per second. Processing total of 103,279,020 data points.
195.87 seconds at 527k data points per second. Processing total of 103,302,428 data points.
186.96 seconds at 553k data points per second. Processing total of 103,302,428 data points.
```
```
PR
87.29 seconds at 910k data points per second. Processing total of 79,452,428 data points.
103.70 seconds at 766k data points per second. Processing total of 79,452,428 data points.
89.56 seconds at 887k data points per second. Processing total of 79,452,428 data points.
```
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/8023
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
N/A
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Existing tests and new test reproducing issue
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [ ] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->